+++ src/Items/Utilities/Evidence.java
@@ -15,15 +15,15 @@ import java.io.Serializable;
  * @since   1.0 - (pre-release version)
  */
 public class Evidence implements Serializable {
-    private int amount;
-    private int price;
+    private long amount;
+    private long price;
 
-    public Evidence(int amount, int price) throws WrongEvidenceException {
+    public Evidence(int amount, long price) throws WrongEvidenceException {
         setAmount(amount);
         setPrice(price);
     }
 
-    public void setAmount(int amount) throws WrongEvidenceException {
+    public void setAmount(long amount) throws WrongEvidenceException {
         if (amount > 0) {
             this.amount = amount;
         } else {
@@ -31,7 +31,7 @@ public class Evidence implements Serializable {
         }
     }
 
-    public void setPrice(int price) throws WrongEvidenceException {
+    public void setPrice(long price) throws WrongEvidenceException {
         if (price > 0) {
             this.price = price;
         } else {
@@ -39,11 +39,11 @@ public class Evidence implements Serializable {
         }
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
@@ -75,17 +75,17 @@ public class Evidence implements Serializable {
      *     <li>{@code Index 3:} stands for how many products were backtracked</li>
      * </ul>
      */
-    public int[] register(int move) {
+    public long[] register(long move) {
         if (move >= this.amount) {
-            int soldAmount = this.amount;
+            long soldAmount = this.amount;
             move = move - this.amount;
-            int oldMoney = this.amount * this.price;
+            long oldMoney = this.amount * this.price;
             this.amount = 0;
-            return new int[]{move, this.amount, oldMoney, soldAmount};
+            return new long[]{move, this.amount, oldMoney, soldAmount};
         } else {
             this.amount = this.amount - move;
-            int oldMoney = move * this.price;
-            return new int[]{0, this.amount, oldMoney, move};
+            long oldMoney = move * this.price;
+            return new long[]{0, this.amount, oldMoney, move};
         }
     }
 }
